thriftrw: Update fx plugin
This updates the ThriftRW plugin used by UberFx to match the most recent
YARPC dev version of the plugin but using UberFx values as defaults for the
command line flags.

This plugin can be completely removed when the next version of YARPC Go is
released.

With this new version, `WrapUnary` and `WrapOneway` no longer accept
`service.Host` as an argument.