Integrate richgo for better test output
A few things are going on here:

First, we're upgrading overalls for a PR I merged this morning that adds
suppor for richgo (https://github.com/kyoh86/richgo). We're currently
pointing at my fork of richgo until my PRs get merged.

Mostly we need this: https://github.com/kyoh86/richgo/pull/6

But I also fixed some stuff about TTY handling that I have an open issue
with the author for.

The second thing I changed is that the coverage data is now dumped to
coverage.txt by default. It grew large enough that it didn't fit on a
single page, or even three pages. The output of richgo actually gives
you package-level coverage and visualization so I think this is a good
tradeoff. A few new ways to run the tests, but the default is my
favorite:

```
make test TEST_FLAGS=-v
```

This shows off richgo's really neat features, but is more verbose

```
make test COVER=1
```

This returns the old behavior of showing the gigantic coverage breakdown
on stdout after each test run.