Restore dev logging functionality
A couple of things was going on here after the large zap refactor. The
most important one is the populateStruct error was not the only
condition to use the default logger, the other is on logging
configuration at all.

Logging setup now returns an error if it's not able to properly parse
the config.