dig: Constructors can now return errors
This allows constructors to fail with an error.

    func NewMyResult(...) (*MyResult, error) {
        // ..
    }

The error is wrapped for every layer so that you get,

    unable to resolve Foo:
      unable to resolve *MyResult:
        root cause

This works by changing `funcNode`s to always be functions in the form,

    func(...) ($ResultType, error)

When constructors are initially registered, if they are in the no-error
form, we wrap them to always return a `nil` error.