Add task.MustRegister which panics on error
Although panics are generally frowned upon, there are certain cases
where they can make developers' lives easier. Particularly for the task
module, the general use case is that in the `main` func (or some other
func called by `main` during initialization), there is no way to recover
from the error and we want to crash the process.

In abscence of MustRegister, a developer who wishes to register many
tasks must check the error of each, and upon encountering an error will
almost certainly want to exit the program with a non-zero exit code
since the service should not start if the task functions are not
properly defined.

In other words, it turns this:

  if err := task.Register(FirstTaskFunc); err != nil {
    log.Fatal("Could not register task", err.Error())
  }
  if err := task.Register(SecondTaskFunc); err != nil {
    log.Fatal("Could not register task", err.Error())
  }

Into this:

  task.MustRegister(FirstTaskFunc)
  task.MustRegister(SecondTaskFunc)

Both of which behave semantically similar (although the output may be
slightly different).

This gives developers the flexibility to choose which behavior they
want, similar to `regexp.MustCompile` from the standard library, where a
program should not start if the developer incorrectly constructed a
regular expression.