Add dotted keys support to yaml provider (#477)
We split the path to a node based on dots, which makes impossible to work with e.g. maps with keys containing dots.

In this change we recursively try to find the longest key that matches dotted path to a node. For example, if we want to find a key a.b in a map["a":map["b":"c"], "a.b":"d"] we want to return d instead of c.