Add fx.Populate to pull values from the container
This adds an `fx.Populate(target)` function, the output of which may be
passed to `App.Start()` to fill the target struct with values from the
DI container.

    var x struct {
        Dispatcher *yarpc.Dispatcher
    }

    app.Start(ctx, Populate(&x))

This enables something very simple that users will expect to be able to
do: start the application, do things with the objects inside the started
application, and then stop it--all in a serial manner rather than
out-of-band in a separate goroutine. Need for this has already become
apparent while writing integration tests that interact with Fx.
See #500 for more discussion.

This works by generating a function accepting a single struct as an
argument. The struct embeds `dig.In` and contains all exported fields
from the user-provided target struct. This lets us delegate most of the
work to dig and allows us to support dig tags like `optional:"true"`
(and possibly `named:".."`). The generated function copies values from
the received struct back to the user-provided struct while leaving
unexported fields alone.