Go 1.9 support: Attempt 3; Drop dig.In support
We now implement this functionality by generating a function which
accepts one argument for each field and fills those back into the
original struct.

A caveat of this is that the structs are no longer implicitly `dig.In`
structs. If users want to use `dig.In` functionality, they will have to
have a field which is a dig.In struct.

    type Foo struct {
        Bar struct {
            dig.In

            Reader io.Reader `optional:"true"`
        }
    }