Top-level API, options, and composable stacks (#514)
This PR proposes a tweak of the top-level application API that makes
stacks much nicer.

The meat of the change is simple: it makes constructors and invokes
options passed to `New`. This has a few benefits:

- Conceptually, it fully defines the application during construction.
This reduces the temptation to pass the application object out of main.
- It lets us mirror dig's `Provide` and `Invoke` terminology, which
makes it easier to navigate docs.
- It naturally leads to the `Options` constructor, which lets a stack
define both provides and invokes in a single object.

You can see an example application in the included example. A stack
might look like this:

```
package fxstack

var Stack = fx.Options(
  fx.Provide(
    foo.Module,
    bar.Module,
    baz.Module,
  ),
  fx.Invoke(
    foo.Run,
    baz.Run,
  ),
)
```

In a user's main function, we'd see this:

```
app := fx.New(
  fxstack.Stack,
  fx.Provide(
    // User-supplied types.
    repository.NewUserRepository,
    gateway.NewTripGateway,
    handler.NewHandler,
  ),
)
app.Run()
```

To reduce the wordiness of users' code, this PR also provides a small
context construction helper and makes `Inject` a standalone `Option`.