Clean up hanging functions in fxlog
The internal fxlog package is structured around an interface that (a)
isn't satisfied by any other logging library, and (b) has only one
implementation. This PR simplifies things a bit by dropping the
interface and making the helper functions into methods. Along the way,
it rationalizes the pointer/value semantics of the logger type - `New`
returns a pointer, and all the methods have pointer receivers.

If we want to make fx's logging pluggable in the future, we'll still
need these changes: we'll want the user to supply an object that
satisfies the same interface as the standard library's `*log.Logger`,
and we'll wrap it in the same concrete type that we're using now.