[4/5] Make Travis builds faster (#546)
* Make Travis builds faster

We don't need to run in VMs, so we'll get some speed improvements by
setting `sudo:false` and using their container-based infra instead. We
should also cache the vendor directory and check licensing in CI.

* Lint already runs license check

* Clean up from CR

* Gimme doesn't know what to do with 1.9.x

* Remove Overcommit (#547)

To be friendly to open-source contributors, let's drop Overcommit. We
lose a little linting power, but we run all the important checks in CI.
Personally, I found it odd that I needed a system-global Ruby gem
installed to work on a Go project.