[2/5] Fix quoting and escaping in check_license.sh (#544)
* Fix quoting and escaping in check_license.sh

Array access is (apparently) supposed to be double-quoted, and escaping
this particular forward-slash wasn't required.

I dunno man, I just do what `shellcheck` tells me :/

* [3/5] Remove extraneous grep from Makefile (#545)

* Remove extraneous grep from Makefile

Since we don't ship a big examples directory anymore, we don't need
this. In a slimmed-down 1.0 world, we should ship examples as GoDoc.

* [4/5] Make Travis builds faster (#546)

* Make Travis builds faster

We don't need to run in VMs, so we'll get some speed improvements by
setting `sudo:false` and using their container-based infra instead. We
should also cache the vendor directory and check licensing in CI.

* Lint already runs license check

* Clean up from CR

* Gimme doesn't know what to do with 1.9.x

* Remove Overcommit (#547)

To be friendly to open-source contributors, let's drop Overcommit. We
lose a little linting power, but we run all the important checks in CI.
Personally, I found it odd that I needed a system-global Ruby gem
installed to work on a Go project.