[discussion] Prototype for running Invokes as part of new
This builds on top of a lot of previous work, most notably the refactor
of the top level API to no longer expose `Provide` and `Invoke` on the
app object. Now that all options have to atomically be provided on
`New`, it no longer makes sense to wait until the application start to
resolve.

I think in order to fully reap the benefits of this change, `app.New`
should return `(*App, err)` tuple.

All the heavy-weight work, like listening for the network, will still be
done as part of the `onStart` hooks.

Lets use this in combination with #550 for a discussion.

Essentially this introduces a new touch point: when the graph is fully
resolved, but the start Hooks have not yet run. I see some potential
downsides of this, such as metrics collectors not being started and
someone trying to interact with them...