Improve the dig.Out provide logging
This PR expands the dig.Out types in the logs to provide a much better
logging experience.

:feelsbadman: Unfortunately, I had to re-implement a bunch of things
about the dig internals: naming conventions for optionals and named
types, and how to parse struct tags. I added a TODO that expands on
another TODO talking about exporting some of the dig internals. Probably
a topic for a conversation. I'd like to land this as is and then think
through the clean up and how dig APIs can make this experience nicer.
Any change would be pretty radical and affect `Provide` and possibly
even `Invoke`.

On the plus side, the logging is now much nicer.

As a recap: `~` means optional, `type:foo` means type named foo. Can
also be combined, see tests.