Fix build against tip
This fixes the build failure against Go tip seen in #601.

The reason for this is that (for some reason), when you run, the tests
with `-covermode=atomic`, the path to the file in the `runtime.Frame` is
*just* the name of the file running the tests rather than the full
absolute path (which contains the string "go.uber.org/fx").

This causes the tests for fxreflect.Caller to fail because fxreflect.go
is no longer considered an "internal" file (since its path doesn't
contain "go.uber.org/fx").

The fix for this is to make sure we skip the fxreflect.go:Caller frame
when fetching the list of callers.