Populate: Support populating any value without requiring structs
Extract requires users to create a struct, but this struct does not
match the standard fx.In struct that users use when using parameters.

Populate will allow users to get values from the dependency injection
container without creating an additional struct, and expects users to
use fx.In for getting multiple values in a struct or if they want to
get named values.