Make it easier to set custom Start and Stop timeouts (#616)
@vkuzmin-uber points out that Fx offers only two options for setting app start/stop timeouts: use the defaults or drop down to the Start/Stop/Done APIs directly. This PR offers a middle ground by introducing StartTimeout and StopTimeout options.

We discussed this quite a bit while designing the Fx 1.x APIs, and agreed to defer making a decision - we wanted a low-level API that let applications use contexts partially consumed by other libraries, but also reasonable behavior by default. Since it turns out that many internal applications need longer to start up but are otherwise happy with Run, introducing these options makes things a bit simpler for them.

As a nice bonus, this makes it possible for fxtest to respect the configured timeouts.