Support visualizing errors with an ErrorHook option (#630)
This change includes changes to fx to use the new features in the
`feature/ graphviz` branch in `uber-go/dig`. On `Invoke` failures, we
pass the error to the registered error handlers provided by users
through `fx.ErrorHook`. This allows users to get a graph representation
on what's causing the invoke failure. The error handlers can then
process the information accordingly (eg.  print to log or write to
file).

* Updated `glide.yaml` and `glide.lock` to use `go.uber.org/dig`
  feature branch.
* Added `ErrorHook` option to `App` to register error handlers.
* `ErrorHandler`s handle the error produced on invoke failure. The
  error is wrapped in a wrapper that also contains the DOT format graph
  generated by `dig.Visualize`.