Annotated: Rename Fn to Target
This renames the Fn field to Target to account for the possibility of
wanting to provide things besides functions with `fx.Provide` and
`fx.Annotated`.

This also alters the docs for Annotated to account for value group
support in the future by decoupling them from the idea of named values.