Drop closure-based fx.Option implementation
We have a handful of `fx.Option` implementations based on `optionFunc`
which captures value using a closure. That makes them opaque and
uninspectable.

This change drops the `optionFunc` type in favor of concrete
implementations for all `fx.Option`s. For further inspection, each such
instance includes a `String()` implementation.

Note that we've intentionally kept `fmt.Stringer` out of the `fx.Option`
interface: We don't want to commit to having options printable but it's
nicer if they are.