fxreflect: Refactor caller name extraction
This pulls out the logic for frame extraction from `fxreflect.Caller`
into its own `func CallStack() Stack` which fills a `Frame` type we own
with sanitized information from `runtime.Frame`s

The `Stack` type supports being formatted with `%v` and `%+v`. With
`%v`, it provides a single-line output and with `%+v`, it provides a
multi-line output in the form,

    foo/bar.Baz()
    	foo/bar/baz.go:42
    baz/qux.Quux()
    	baz/qux/quux.go:50

This is the same format Go uses for stack traces.

This change is meant to ease upcoming debuggability work (T4487249).