fxreflect: Improve frame skipping
fxreflect's frame skipping logic is used to find user-owned functions
without hard-coding skips for Fx functions.

Previously, it did so by hard-coding based on file paths. With Go
modules, we're no longer guaranteed to be in a path that contains
go.uber.org/fx, so instead we will use the fully-qualified function name
to determine if we need to skip.

The fully-qualified function name takes form similar to the following,

    go.uber.org/fx.Populate
    go.uber.org/fx_test.TestFoo
    go.uber.org/fx/internal/lifecycle.(*Lifecycle).Append()

So we can filter out Fx-defined functions by matching on the Fx import
path in the fully-qualified function name.