Add invoke sorting
Invoke ordering has been a huge problem in the last years
that Fx has been live. Executing invokes in the order they
are provided to the fx.New() call is simply not good enough
as it often requires developers to think through the import
order of their modules, and how they interact with one
another.

Luckily help is at hand with a new fx.Option -- fx.Sort! Now
developers can finally take control of their own destiny
when it comes to the order of execution of invokes with the
two very deterministic launch orders: alphabetical, and by
length of the fully qualified function name (comes out of
the box with package name support!!!).

I think this is going to take the possibilities through the
roof.

The framework is there for adding more sorting options in
the future. Some ideas that have been thrown out:
- Random
- Reverse
- Probability
- Time of day
- Leap year

The possibilities are literally endless.