In/Out: Use type aliases
When `fx.In` and `fx.Out` were originally implemented on top of `dig.In`
and `dig.Out`, type aliases didn't exist so they were implemented by
embedding the structs.

Now that we're well past Go 1.9 (when type aliases were added), we
should be more than safe to switch over to type aliases.

Note that even though we're changing the definition of the type, this is
not a breaking change because the `dig.In` and `dig.Out` types were not
usable directly before either as they embedded an internal sentinel
type.

Resolves #730
Refs GO-374