fxtest.NewLifecycle: Make testing.T optional (#742)
With #721, we started writing the output of fxtest.Lifecycle to the
testing.T rather than stderr. However, after experimenting with this
internally, we found that many users supply a `nil` testing.T to
fxtest.Lifecycle in some cases. With just #721, this will begin
panicking.

Resolve this by making testing.T an optional argument of
fxtest.Lifecycle. If unspecified, we'll retain the old behavior of
printing to stderr.

As an added usability bonus, if RequireStart or RequireStop fail,
instead of panicking with a nil deref (which will happen even without
the change from #721 if the testing.T is nil), panic with a more
meaningful message that will help users discern why this failed.

Refs GO-631