tools: Add internal linter, lint fxevent.Loggers
With #733, we have two fxevent.Logger implementations in Fx, and we plan
to change the event types a fair bit in the near future. Rather than
manually verify that the two linters are in-sync and handle all event
types, it's preferable to verify this automatically.

Add a custom linter based on go/analysis meant for use only inside Fx.
This linter verifies that any fxevent.Logger implementations it finds
handle all fxevent.Event types, or none of them.

For testing with analysistest, this creates a fake fxevent package
similar to the real fxevent package, and lints code that uses it.