Add WithLogger option to supply a fxevent.Logger (#733)
This is a follow up to #717. This change adds a `WithLogger` option that
accepts a constructor that returns an `fxevent.Logger`. Unlike
`fx.Printer`, this constructor can have dependencies on other types
provided to the Fx graph. If supplied, we'll use this logger to log
events rather than our default behavior.

To make this work, we do the following:

- If `WithLogger` was not supplied, we use the old behavior of logging
  to stderr or the user-supplied `fx.Printer`.
- If `WithLogger` was supplied, we buffer log messages in-memory until
  we can build the user-supplied logger. If we successfully build the
  logger, we flush the buffered messages to it, and log to it from that
  point onwards. If we were unable to build the logger, we flush the
  buffered messages to stderr, or the user-supplied `fx.Printer` and
  fail application startup.

Further, to match our old behavior more accurately, this change includes
a `ConsoleLogger` implementation of `fxevent.Logger` that behaves almost
exactly like our logger before #717. This will make it easy to roll this
feature out without affecting users negatively, and also give us the
option of selecting between readable and non-readable log outputs based
on runtime environment (production vs development, for example).

Ref GO-77

Co-authored-by: Abhinav Gupta <abg@uber.com>