Add events for lifecycle hook execution (#759)
This adds events for lifecycle hook execution. LifecycleHookStart and LifecycleHookStop was being used to fire events separately for OnStart and OnStop events previously. Since the recorded fields for both events aren't different, this change merges the two events into a single set of events, and adds another event so that an event is fired before and another is fired after a hook is executed.

Specifically, LifecycleHookExecuting is fired right before either an OnStart or OnStop hook is executed. After the hook has finished executing, LifecycleHookExecuted is fired with either an error or the runtime. Both LifecycleHookExecuting and LifecycleHookExecuted events contain the name of the hook being executed, the name of the caller that appended the lifecycle, and the method that's being executed (either OnStart or OnStop).

Refs: GO-707.