LifecycleHookExecuted: Instantiate once each for OnStart/Stop
The way the code for emitting LifecycleHookExecuted was structured, we
would instantiate it separately for OnStart successes and errors, and again
for OnStop successes and failures.

To prevent issues in the future, I feel that it's cleaner to instantiate it
once per case.

Restructure lifecycle execution code to separate hook
execution-and-emission from internal state tracking.