Test that timeouts during rollback are also propagated
If a service fails to start up because a hook failed,
we rollback the hooks that we did succeed at starting up.
If stopping a hook takes too long, that also causes a timeout.
Test that if such an OnStop hook takes too long,
the timeout event is still emitted.