fxevent: Log Started/Stopped on timeout (#789)
As reported in #782, `App.Run` had a bug so that
if an application using it failed to start up because of a timeout
no event would be logged to indicate this.
The system would silently exit with a non-zero status code.

This PR resolves this by ensuring that even for timeouts,
we log `Started` or `Stopped` events with non-nil `Err` values.
This matches the behavior for other startup or shutdown failures.

To test this, I wanted to use `App.Run` directly so
I added an internal test-only option `WithExit`
that lets us override the behavior of `os.Exit`
without a global variable. (#792)

The issue resolution was pretty straightforward:

- Start: `defer LogEvent(..)` in Start instead of
  the wrapped `start` method which runs inside a separate gorotuine.
- Stop: Apply `defer LogEvent(..)` in Stop instead of `App.Run`.
  This actually exposed a bug too: we did not previously log `Stopped`
  if `App.Stop` was called. This resulted in some test changes.

There was a small non-determinism in `withTimeout`:
What if the context times out at the same time as the callback finishing?
`select` will pick one of the `case`s randomly.
I resolved this non-determinism by checking `ctx.Err()` in the success case again.

I also moved `App.start` next to `App.Start`, and `App.run` next to `App.Run`
to aid in readability. The move is in #791.

Resolves #782
Resolves GO-866