Fix panic on Annotate non-nil error returns
If a non-nil error gets returned by a constructor being annotated with
result tags, it panics with an error due to a bug in the reflection code
that aggregates and assigns the error value to the result struct.

The bug is specifically with calling `Elem()` which casts the type to the
element of the error, not the error type itself, resulting in a type mismatch
during reflect.MakeFunc.