CI: Fix Windows FIND warning
Currently, CI on Windows reports the following error:

```
FIND: Parameter format not correct
```

This is because of our eager evaluation of the `find` commands above
that are not quite valid on Windows.

Replace one unnecessary `find` command,
and turn the other to lazy evaluation.
With `=` instead of `:=`, that `find` will be evaluated in the
`make lint` task only, which doesn't run on Windows.