Allow multiple As Annotations to be provided (#817)
Currently, Fx errors out when two As annotations are provided to a constructor. This is inconvenient as there may be times where a user wants to annotate a single type as many interfaces that it implements.

This change allows such annotations to be provided.

In particular, the following annotation is now possible:

fx.Annotate(func() *bytes.Buffer {...},
  fx.As(new(io.Writer)),
  fx.As(new(io.Reader)),
}

bytes.Buffer implements both io.Writer and io.Reader, hence this provides the same instance of bytes.Buffer as both io.Writer and io.Reader interface. Note that this provides the same instances as arguments that takes in these, and should NOT be used as if the constructor got invoked twice to provide two different instances of type bytes.Buffer, one as io.Writer and the other as io.Writer.

To allow this change, few modifications were made to internal representation of As tags. Most of the changes are changing the dimensions of the slices from 1D to 2D. The following changes were made:

annotated.As is now an array of array of reflect.Type, where As[i][j] is the i-th type that j-th result needs to be annotated as.
offsets and results are now 2D slices as well, similar to annotated.As.
The out type of annotated function generated by reflection is now multiple fx.Out structs, each of which correspond to one line of fx.As annotation, followed by an error type if one exists in the output type of the annotated constructor.
Refs: #811
GO-1023