Add more Decorate tests
This adds more cases of fx.Decorate tests, specifically for
transitive dependencies that are decorated. For example, when a
function depends on type A whose constructor depends on type B,
and B has a decorator associated with it, we expect type A's constructor
to use the decorated value of B in its constructor.