Emit a separate log for fx.Replace
This adds a new type fxevent.Replaced, which is used for logging
values that are being decorated via fx.Replace. Currently, these
are being emitted as part of the fxevent.Decorated log, but that
can be misleading and confusing since the decorator name for such
values is not properly provided (they are generated at runtime using
reflection).

The new log entry for fx.Replace-d types contains the types being
replaced, as well as the module name (if there is one).