Fix documentation for fxevent/Decorated and fxevent/Replaced
The current documentation fox fxevent/Decorated is misleading because
it says that it's emitted when a decorator is provided to the fx.App.

In fact, it is emitted when the decorator actually runs in the fx.App.
The current doc is misleading, so this fixes the wording. Also, this
fixes the description for the fxevent/Replace event as well to be more
specific to Replace.