Add machinery for verified docs
This adds a new docs/ folder to the Fx repository.
This is where non-API-reference documentation will reside.

To set things up, this adds a docs/CONTRIBUTING.md
with guidance on how to contribute to the Fx documentation.

It includes tooling to ensure that code snippets
inside documentation remain up to date.
To do that, this relies on two tools:

- mdox does the bulk of the heavy lifting here
- the region shell script helps pick out named snippets of code
  from a flie