ci/docs: Fix setup-node, install, yarn registry
The setup-node action expects the yarn.lock to be present
in the root of the repository,
or if it's in a subdirectory, for a path to be specified.

https://github.com/actions/setup-node/blob/main/docs/advanced-usage.md#caching-packages-data

Further, yarn adds links to the currently used yarn registry to the
yarn.lock.
In our case, it was pointing to the internal registry,
so I've added project-local configuration to use the global yarn
registry.

Lastly, we were running yarn install in the root of the repository,
not the docs subdirectory.