docs: Add release notes
Include release notes on our docs website
by adding a symlink of CHANGELOG.md to the docs folder.

Because this page is long and can get noisy,
it opts out of being indexed for search,
and doesn't show the list of headers in the sidebar.

Finally, this reformats the changelog with mdox.