Error When Starting Lifecycle Twice
Currently, it is possible for a user to start their app (and thus
its lifecycle) twice. This can lead to panics if the user has
registered any lifecycle hooks (Ref:
https://github.com/uber-go/fx/issues/991).

This change will explicitly keep track of whether a lifecycle
is currently running, under a lock, to ensure that the lifecycle
can never be running twice at the same time.