Stricter Control over Starting/Stopping Lifecycle (#994)
Currently, there is nothing preventing somebody from making calls to
`app.Start` and `app.Stop` when it doesn't make sense, for example
calling `app.Start` or `app.Stop` twice, or trying to run the
application twice concurrently. Because of this lack of control, it's
possible for users to cause race conditions and panics (#991) by putting
the lifecycle is strange positions.

This PR places a small state machine in `Lifecycle` so that we can check
if calls to `(*Lifecycle).Start` and `(*Lifecycle).Stop` make sense
before going through. This will disallow starting an already
starting/started app, and stopping an already stopping/stopped app.
Tests to verify that are also added.