docs: Disable Go reformatting (#1000)
(This is a new version of #961. I can't push to the original branch
anymore.)

---

Upgrade to latest markdownfmt
which does not reformat Go source code blocks by default.

With this change, formatting of code blocks in Fx docs will be
consistent
instead of some being gofmt-ed to use tabs,
and others using 2 spaces (per our 'region' script).

This is a landable version of #961 without any temporary version pins.