Remove extraneous error logs in lifecycle deadline exceed errors
When running lifecycle hooks, it's possible they fail due to context
deadline exceeded errors. Fx currently reports this error with a
sorted list of all the hooks that have ran. This was added because
it was difficult to debug which hook caused the context deadline
error to occur.

Now that every hook that is run emits an associated fxevent (
OnStartExecuting, OnStartExecuted, OnStopExecuting, OnStartExecuted),
this error wrapping is no longer necessary.

Fixes #999