Allow fx.Annotate on In/Out structs with ResultTags/ParamTags, respectively.
It is currently not possible for fx.Annotate to be used on a function that
takes in or returns fx.In and fx.Out structs.

This limitation, however, is not necessary when the annotation is different
from the struct defined annotations, such as using fx.In with fx.ResultTags
Annotation or fx.Out with fx.ParamTags Annotation.

This loosens the constraint so that these Annotations become valid.

Fixes #980.