Fix race test for Start/Stop hooks
The race test that was added in Start/Stop hooks in #1062. This test
can fail if the machine is sufficiently slow that the second Stop
hook does not finish running before the test finishes.

This removes the check for the second hook from running - this is still
adequately testing the code path we wanted to test since the race
would occur on the first stop hook with the second start hook.