ci: Add an internal/e2e submodule
Adds an internal/e2e submodule to place end-to-end tests in.
These can be full-fledged Fx applications that we run tests against.

This is a submodule so that it can have dependencies that are otherwise
not desirable for Fx.
This is inside the internal/ directory so that it can consume
Fx-internal packages (despite being in a separate submodule).