Make fx.Populate work with fx.Annotate (#1089)
This PR makes fx.Populate work with fx.Annotate. Previously, to
fx.Populate a named dependency, users would have to create a 
struct that embeds fx.In and has a field with a tag and pass a 
pointer to that struct. This could be cumbersome in cases where 
the user is intersted in only one named value. With this PR,
users will be able to use fx.Annotate to get the same effect:

```
var a A
...
fx.Populate(
  fx.Annotate(
    &a,
    fx.ParamTags(`name:"A"`),
  ),
)
```