app.go: Take a pass over the godoc
The godocs for most of the public API in app.go hasn't been touched in a
while.
I took a quick pass over it to adjust it better to current
recommendations and best practices.
This also makes more use of Go's `[..]` linking.

One notable change: Instead of Options, logical groups of options are
now recomended via Module.